Request form locations, filename changes, basic styling
