+++ nbactions.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <actions>
     <action>
-       <actionName>CUSTOM-appengine:run</actionName>
+        <actionName>CUSTOM-appengine:run</actionName>
         <displayName>appengine:run</displayName>
         <goals>
             <goal>appengine:run</goal>
