+++ src/main/java/com/google/step/QueryDrive.java
@@ -64,6 +64,4 @@ public class QueryDrive extends HttpServlet {
             });
         }
     }
-
-
 }
