+++ src/main/java/com/google/step/SaveMapImageDatastore.java
@@ -1,58 +1,36 @@
- package com.google.step;
+package com.google.step;
 
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
-import com.google.cloud.storage.Blob.BlobSourceOption;
-import com.google.cloud.storage.Blob;
-import com.google.cloud.storage.BlobId;
-import com.google.cloud.storage.BlobInfo;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageException;
-import com.google.cloud.storage.StorageOptions;
-import java.awt.image.BufferedImage;
 import java.io.BufferedReader;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.net.URL;
-import java.util.Date;
-import java.util.List;
-import java.util.ArrayList;
 import javax.imageio.ImageIO;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import com.google.api.services.drive.Drive;
-import com.google.api.services.drive.DriveScopes;
-import com.google.api.client.json.jackson2.JacksonFactory;
-import com.google.api.client.extensions.appengine.http.UrlFetchTransport;
-import com.google.api.client.googleapis.extensions.appengine.auth.oauth2.AppIdentityCredential;
-import com.google.api.services.drive.Drive.Files;
-import com.google.api.services.drive.model.File;
-import com.google.api.services.drive.model.FileList;
 import java.util.Arrays;
 import java.util.ArrayList;
-import java.util.List;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
-// TODO: Tests
-
-@WebServlet("/save-datastore")
+/***
+    This servlet retrieves MapImages after many attributes are set in SaveImageCloud.java to be stored in Datastore.
+    A POST request gets the MapImage ArrayList to store each MapImage instance in Datastore.
+***/
+@WebServlet("/save-mapimage-datastore")
 public class SaveMapImageDatastore extends HttpServlet {
 
-    /** Stores updated MapImage object metadata into Datastore. */
+    /** Stores the updated MapImage objects' metadata into Datastore. */
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        
-        // Get  MapImage objects here from SaveImages.java
+        // Get  MapImage objects here from SaveImageCloud.java
         BufferedReader reader = request.getReader();
         Gson gson = new Gson();
         ArrayList<MapImage> mapImages = gson.fromJson(reader, new TypeToken<ArrayList<MapImage>>(){}.getType());  
 
-        // Put updated mapImages into Datastore of entity kind MapImage. Entity look-up key is objectID.
+        // Put updated MapImages into Datastore under the entity kind 'MapImage'. An entity's look-up key is its objectID.
         DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
         for(MapImage mapImage : mapImages) {
             Entity mapImageEntity = new Entity("MapImage", mapImage.getObjectID());
@@ -67,5 +45,4 @@ public class SaveMapImageDatastore extends HttpServlet {
             datastore.put(mapImageEntity);
         }
     }
-
 }
