+++ src/main/webapp/WEB-INF/cron.yaml
@@ -3,7 +3,7 @@
 ---
 cron:
 - description: "monthly map snapshots"
-  url: /query-datastore
+  url: /backend-query-datastore
   schedule: 1 of month 00:00
   timezone: America/Seattle
   retry_parameters:
