+++ src/main/webapp/WEB-INF/logging.properties
@@ -1,4 +1,3 @@
-
 # A default java.util.logging configuration.
 # (All App Engine logging is through java.util.logging by default).
 #
