+++ src/main/webapp/app.html
@@ -10,7 +10,6 @@
     <script src="script.js"></script>
 </head>
 <body>
-
     <div id="main-container">
  
         <h1>Maps Over Time</h1>
@@ -48,8 +47,8 @@
             <option value="8">8</option>
             <option value="9">9</option>
             <option value="10">10</option>
-            <option value="11">11</option>            
-            <option value="12">12</option>            
+            <option value="11">11</option>
+            <option value="12">12</option>
             </select>
 
             <br>
