+++ src/main/webapp/script.js
@@ -1,6 +1,7 @@
 $(document).ready(function() {
     fetch("/query-cloud").then(response => response.json()).then(array => {
-        // Array is not equal to "{}" when the request is made after the form is submitted
+        // array will be "{}" since a request is made before the form is submitted (on page load).
+        // array will not equal "{}" when request is made after form submission.
         if (array !== "{}") {
             for (var i = 0; i < array.length; i++) {
                 const url = array[i].url;
@@ -13,20 +14,28 @@ $(document).ready(function() {
     fetch("/form-locations").then(response => response.json()).then(locations => {
         $("#locations").empty();
         var emptyOption = $('<option></option>').attr("value", "").text("");
-        $("#locations").append(option);
+        $("#locations").append(emptyOption);
         for (var j = 0; j < locations.length; j++) {
             var option = $('<option></option>').attr("value", locations[j]).text(locations[j]);
             $("#locations").append(option);
         }
     });
-
 })
 
+/** Removes any current description and image li elements on the page. */
+async function clearImages() {    
+    var list = document.getElementById("requestedImages");
+    // As long as <ul> has a child node, remove it
+    while (list.hasChildNodes()) {
+        list.removeChild(list.firstChild);
+    }
+}
 
+// CODE BELOW is not incorporated currently, but is being referenced while writing jQuery above.
 
 /** Builds Unordered List of map snapshots and their descriptions. */
 function getImages() {
-    fetch('/query-drive').then(response => response.json()).then((listMapImages) => {       
+    fetch('/query-drive').then(response => response.json()).then((listMapImages) => {
         // listMapImages is an ArrayList of MapImages with their metadata which includes a url attribute for image data.
         const imageDisplay = document.getElementById('requestedImages');
         listMapImages.forEach((mapImage) => {
@@ -51,15 +60,4 @@ function createListImage(url) {
     imgElement.setAttribute("src", url);
     liElement.appendChild(imgElement)
     return liElement;
-}
-
-
-/** Removes any current description and image li elements on the page. */
-async function clearImages() {    
-    var list = document.getElementById("requestedImages");
-    
-    // As long as <ul> has a child node, remove it
-    while (list.hasChildNodes()) {
-        list.removeChild(list.firstChild);
-    }
 }
